# [Fix] - Router: add model_name index for O(1) deployment lookups