## Title

[Fix] - Router: add model_name index for O(1) deployment lookups
<!-- e.g. "Implement user authentication feature" -->

## Relevant issues
## Pre-Submission checklist

**Please complete all items before asking a LiteLLM maintainer to review your PR**

- [ ] I have Added testing in the [`tests/litellm/`](https://github.com/BerriAI/litellm/tree/main/tests/litellm) directory, **Adding at least 1 test is a hard requirement** - [see details](https://docs.litellm.ai/docs/extras/contributing_code)
- [ ] I have added a screenshot of my new test passing locally 
- [x] My PR passes all unit tests on [`make test-unit`](https://docs.litellm.ai/docs/extras/contributing_code)
- [x] My PR's scope is as isolated as possible, it only solves 1 specific problem


## Type

🧹 Refactoring

## Changes
- Add model_name_to_deployment_indices index for O(1) model name lookups
- Optimize _get_all_deployments() from O(n) scan to O(1) + O(k) index lookup
- Maintain both model_id and model_name indices in _add_model_to_list_and_index_map()

## Performance Gains
<p><strong>Summary</strong></p>
<p>Average, min and rps changed the most.</p>
<p><strong>Before</strong></p>

Type | Name | # Requests | # Fails | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Min (ms) | Max (ms) | Average size (bytes) | Current RPS | Current Failures/s
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
POST | /chat/completions | 102040 | 2 | 500 | 830 | 1200 | 532.05 | 67 | 2527 | 398 | 778.4 | 0
Custom | LiteLLM Overhead Duration (ms) | 102038 | 0 | 47 | 71 | 97 | 50.42 | 4 | 1438 | 0 | 778.4 | 0
  | Aggregated | 204078 | 2 | 290 | 700 | 1100 | 291.24 | 4 | 2527 | 199 | 1556.8 | 0
  |   |   |   |   |   |   |   |   |   |   |   |  


<!-- notionvc: 715c9aa8-d26f-4126-8bc2-f0ecbeeded95 -->

<p><strong>After</strong></p>

Type | Name | # Requests | # Fails | Median (ms) | 95%ile (ms) | 99%ile (ms) | Average (ms) | Min (ms) | Max (ms) | Average size (bytes) | Current RPS | Current Failures/s
-- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | -- | --
POST | /chat/completions | 115475 | 6 | 450 | 800 | 1300 | 486.82 | 23 | 2606 | 397.98 | 877.9 | 0
Custom | LiteLLM Overhead Duration (ms) | 115469 | 0 | 44 | 73 | 110 | 47.86 | 1 | 1779 | 0 | 877.9 | 0
  | Aggregated | 230944 | 6 | 200 | 660 | 1200 | 267.35 | 1 | 2606 | 198.99 | 1755.8 | 0


<!-- notionvc: 7c1bbb67-aca5-4cfc-af25-5fb2c9a230ad -->